% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categoricalAestheticMap-module.R
\name{add_aesthetic_covariates}
\alias{add_aesthetic_covariates}
\title{Adds aesthetic covariates from a CategoricalAesMapModule to a facile_frame.}
\usage{
add_aesthetic_covariates(aes, to, ...)
}
\arguments{
\item{aes}{the \code{CategoricalAesMapModule}}

\item{to}{the facile_frame to add it to}
}
\description{
Note that this can only happen in a reactive context
}
