% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assaySelect-module.R
\name{assay_feature_info.AssaySelectInput}
\alias{assay_feature_info.AssaySelectInput}
\title{Retrieve the features assocaitd with an assay}
\usage{
\method{assay_feature_info}{AssaySelectInput}(x, assay_name,
  feature_ids = NULL, ...)
}
\arguments{
\item{x}{An \code{AssaySelect} object, returned fom \code{\link[=assaySelect]{assaySelect()}}}
}
\value{
a tibble of features
}
\description{
Retrieve the features assocaitd with an assay
}
