% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assayFeatureSelect-module.R
\name{assayFeatureSelect}
\alias{assayFeatureSelect}
\title{A select input for the user to pick features from an assay.}
\usage{
assayFeatureSelect(input, output, session, rfds, assay = NULL, ...,
  .exclude = NULL, .reactive = TRUE)
}
\arguments{
\item{.exclude}{tibble of feauture_type,feature_id pairs that should not be
included in this selector}
}
\description{
If the user passes in the name of the assay as a reactive in the \code{assay}
parameter (for the module, or as the character string in the UI function),
then this module doesn't provide an "assay" dropdown, and only provides
the features for that assay.
}
