% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assayFeatureSelect2-module.R
\name{assayFeatureSelectServer}
\alias{assayFeatureSelectServer}
\title{A module to pull out features from an assay with optional GeneSetDb support.}
\usage{
assayFeatureSelectServer(
  id,
  rfds,
  gdb = reactive(NULL),
  ...,
  exclude = NULL,
  debug = FALSE
)
}
\description{
A module to pull out features from an assay with optional GeneSetDb support.
}
