% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assaySelect-module.R
\name{assaySelect}
\alias{assaySelect}
\title{A module to create a dropdown over the current assays defined for a datastore}
\usage{
assaySelect(
  input,
  output,
  session,
  rfds,
  ...,
  .exclude = NULL,
  .reactive = TRUE
)
}
\arguments{
\item{rfds}{A ReactiveFacileDataStore}
}
\description{
A module to create a dropdown over the current assays defined for a datastore
}
