% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assaySelect-module.R
\name{assaySelectServer}
\alias{assaySelectServer}
\title{Dropdown widget to select assays over the set of active_samples}
\usage{
assaySelectServer(id, rfds, ..., debug = FALSE)
}
\description{
Dropdown widget to select assays over the set of active_samples
}
