% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCorrectConfig-module.R
\name{batchCorrectConfig}
\alias{batchCorrectConfig}
\title{A module to configure batch correction parameters for assay data.}
\usage{
batchCorrectConfig(
  input,
  output,
  session,
  rfds,
  color = FALSE,
  facet = FALSE,
  group = FALSE,
  hover = FALSE,
  shape = FALSE,
  ...,
  .with_none = TRUE,
  .exclude = NULL,
  .reactive = TRUE
)
}
\description{
This module enables the user to select covariates that should be regressed
out in the data. At the same time a covariate can be selected that indicates
structure in the data to preserve.
}
\examples{
\dontrun{
# ui code
batchCorrectConfigUI(ns("batch"), direction = "horizontal")

# server code
batch <- callModule(batchCorrectConfig, "batch", rfds)
dat <- reactive({
  these_samples() |>
    with_assay_data(these_features(), normalized = TRUE,
                    batch = name(batch$batch),
                    main = name(batch$main))
})
}
}
