% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categoricalAestheticMap-module.R
\name{categoricalAestheticMapServer}
\alias{categoricalAestheticMapServer}
\title{Module to map categorical aesthetics to data.}
\usage{
categoricalAestheticMapServer(
  id,
  rfds,
  color = FALSE,
  shape = FALSE,
  facet = FALSE,
  hover = FALSE,
  group = FALSE,
  ...,
  with_none = TRUE,
  exclude = reactive(NULL),
  debug = FALSE
)
}
\description{
Module to map categorical aesthetics to data.
}
