% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categoricalSampleCovariate-modules.R
\name{categoricalSampleCovariateLevels}
\alias{categoricalSampleCovariateLevels}
\title{Use this with categoricalSampleCovariateSelect to enumerate its levels.}
\usage{
categoricalSampleCovariateLevels(
  input,
  output,
  session,
  rfds,
  covariate,
  missing_sentinel = NULL,
  ...,
  .exclude = NULL,
  .reactive = TRUE,
  debug = FALSE
)
}
\arguments{
\item{missing_sentinel}{This is a reactive (string). When it's NULL, no
missing sentinel is added. The parent covariate selector can pass in a
value here to show to indicate a level that's not included in the
covariate's level.}

\item{covaraite}{the \code{categoricalSampleCovariateSelect} module.}
}
\description{
Use this with categoricalSampleCovariateSelect to enumerate its levels.
}
