% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamodFacileDataStoreServer.R
\name{datamodFacileDataStoreServer}
\alias{datamodFacileDataStoreServer}
\title{A filterable ReactiveFacileDataStore based on the datamods mojo}
\usage{
datamodFacileDataStoreServer(
  id,
  x,
  ...,
  ignore_vars = shiny::reactive(NULL),
  samples_subset = shiny::reactive(NULL),
  user = Sys.getenv("USER"),
  debug = FALSE
)
}
\arguments{
\item{id}{the id of the module}

\item{x}{the thing that will instantiate a FacileDataSet}

\item{user}{the user using it}
}
\value{
a DatamodFacileDataStore,ReactiveFacileDataStore, ...,
}
\description{
A filterable ReactiveFacileDataStore based on the datamods mojo
}
