% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facileBoxplotModule.R
\name{facileBoxplot}
\alias{facileBoxplot}
\alias{facileBoxplotUI}
\title{A shiny fboxplot module}
\usage{
facileBoxplot(input, output, session, rfds, ...)

facileBoxplotUI(id, ...)
}
\description{
Should this go into FacileViz?
}
