% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facileDataSetList-module.R
\name{facileDataSetSelectServer}
\alias{facileDataSetSelectServer}
\title{Lists the FacileDataSet objects in a parent directory for a user to choose.}
\usage{
facileDataSetSelectServer(id, datadir, metafn = "meta.yaml", ...)
}
\arguments{
\item{id}{the ID of the module}

\item{datadir}{the directory that holds the FacileDataSet directories}
}
\value{
A list with reactive components:
\enumerate{
\item \verb{$path()}: The path to the FacileDataSet
\item \verb{$gdb()}: A GeneSetDb object to match the organism of the FDS at \verb{$path}.
}
}
\description{
This module provides a list of FacileDataSet objects to choose from that
live in a parent directory. It will be of most use when building an
application that allows the exploratory data analysis across a number of
different sources of data.
}
\details{
Enumerates the available dataset objects in \code{datadir} once upon invocation.
A \code{meta.yaml} file in \code{datadir} will tell it how to present (name) and group
the datasets from a dropdown select.
}
\section{Data Directory}{

The \code{datadir} must present itself as a locally accessible directory to the
system running this module.
}

