% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamodFacileDataStoreServer.R
\name{FacileDataStoreFactory}
\alias{FacileDataStoreFactory}
\title{Helper function to create a FacileDataStore (really a FacileDataSet).}
\usage{
FacileDataStoreFactory(x, ...)
}
\arguments{
\item{x}{a path to a faciledataset on disk, a faciledataset itself, etc.}

\item{...}{dunno what else}
}
\value{
A FacileDataStore\link{Set}
}
\description{
This function figures out what \code{x} is and turns into into a FacileDataStore.
}
