% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facileFrameFilters-module.R
\name{facileFrameFilterServer}
\alias{facileFrameFilterServer}
\title{Dynamic facile_frame filters that leverage the power of a lazy facile_frame}
\usage{
facileFrameFilterServer(id, x, ..., vars = reactive(NULL))
}
\arguments{
\item{id}{the id of the module}

\item{x}{a reactive(facile_frame)}

\item{vars}{the names of the columns (sample_covariates)}
}
\description{
Dynamic facile_frame filters that leverage the power of a lazy facile_frame
}
