% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facileSampleFilter-module.R
\name{facileSampleFilter}
\alias{facileSampleFilter}
\alias{facileSampleFilterUI}
\title{A composable module to filter to a subset of samples from a FacileDataStore.}
\usage{
facileSampleFilter(input, output, session, rfds, ...)

facileSampleFilterUI(id, ...)
}
\description{
This module will eventually be composed / replicated by adding more or less
filters to restrict (or expand) the sample space to be anlayzed.
}
\details{
I think this means that the set of samples this filter can operate on are
fixed, but covariates can change given interactivity from the user during
an analysis session.
}
