% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facileBoxPlot-module.R,
%   R/facileFrameScatterPlot-module.R, R/facileScatterPlot-module.R
\name{facileBoxPlot}
\alias{facileBoxPlot}
\alias{facileBoxPlotUI}
\alias{facileFrameScatterPlot}
\alias{facileScatterPlot}
\alias{facileScatterPlotUI}
\title{An interactive scatterplot module (2-3 dimensions).}
\usage{
facileBoxPlot(
  input,
  output,
  session,
  rfds,
  ...,
  x = NULL,
  y = NULL,
  event_source = session$ns("selection"),
  .reactive = TRUE
)

facileBoxPlotUI(id, ...)

facileFrameScatterPlot(id, ...)

facileFrameScatterPlot(id, ...)

facileScatterPlot(
  input,
  output,
  session,
  rfds,
  ...,
  ndim = 3,
  x = NULL,
  y = NULL,
  z = NULL,
  event_source = session$ns("selection"),
  .reactive = TRUE
)

facileScatterPlotUI(id, with_download = TRUE, ...)
}
\arguments{
\item{ndim}{Defaults to 3. When any two dimensions are provided, a plot will
be drawn, so provides both 2d and 3d functionality. If set to 2, then only
2d functionality would be enabled.}
}
\description{
This module presents all three dimensions for plotting. If the user only
specifies two dimensions in a weird way (x and z, only), then z will be
collapsed to y.

This module presents all of the numeric columns in x + internal quantitative
data for plotting, and categorical covariates for aesthetic mapping

This module presents all three dimensions for plotting. If the user only
specifies two dimensions in a weird way (x and z, only), then z will be
collapsed to y.
}
\details{
We can make this more user-proof by incrementally activating "the next"
dimension by ensuring the previous dimension is selected.

A 1d scatter plot might just default to a density plot.

We can make this more user-proof by incrementally activating "the next"
dimension by ensuring the previous dimension is selected.

A 1d scatter plot might just default to a density plot.
}
