% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fboxPlot-modules.R
\name{fboxPlotServer}
\alias{fboxPlotServer}
\title{Module server for boxplots}
\usage{
fboxPlotServer(
  id,
  rfds,
  ...,
  gdb = NULL,
  x = NULL,
  y = NULL,
  event_source = session$ns("selection"),
  .reactive = TRUE,
  debug = FALSE
)
}
\description{
Module server for boxplots
}
