% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assaySelect-module.R
\name{features.AssaySelectInput}
\alias{features.AssaySelectInput}
\title{Retrieve the features associatd with an assay}
\usage{
\method{features}{AssaySelectInput}(x, assay_name, feature_ids = NULL, ...)
}
\arguments{
\item{x}{An \code{AssaySelect} object, returned fom \code{\link[=assaySelect]{assaySelect()}}}
}
\value{
a tibble of features
}
\description{
Retrieve the features associatd with an assay
}
