% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facile-datamods-filter-data.R
\name{filter-data}
\alias{filter-data}
\alias{ffilter_data_ui}
\alias{ffilter_data_server}
\title{Shiny module to interactively filter a \code{data.frame}}
\usage{
ffilter_data_ui(id, show_nrow = TRUE, max_height = NULL)

ffilter_data_server(
  id,
  data = reactive(NULL),
  vars = reactive(NULL),
  name = reactive("data"),
  defaults = reactive(NULL),
  drop_ids = getOption("datamods.filter.drop_ids", default = TRUE),
  widget_char = c("virtualSelect", "select", "picker"),
  widget_num = c("slider", "range"),
  widget_date = c("slider", "range"),
  label_na = "NA",
  value_na = TRUE
)
}
\arguments{
\item{id}{Module id. See \code{\link[shiny:moduleServer]{shiny::moduleServer()}}.}

\item{show_nrow}{Show number of filtered rows and total.}

\item{max_height}{Maximum height for filters panel, useful
if you have many variables to filter and limited space.}

\item{data}{\code{\link[shiny:reactive]{shiny::reactive()}} function returning a
\code{data.frame} to filter.}

\item{vars}{\code{\link[shiny:reactive]{shiny::reactive()}} function returning a
\code{character} vector of variables for which to add a filter.
If a named \code{list}, names are used as labels.}

\item{name}{\code{\link[shiny:reactive]{shiny::reactive()}} function returning a
\code{character} string representing \code{data} name, only used for code generated.}

\item{defaults}{\code{\link[shiny:reactive]{shiny::reactive()}} function returning a
named \code{list} of variable:value pairs which will be used to set the filters.}

\item{drop_ids}{Drop columns containing more than 90\% of unique values, or than 50 distinct values.
Use \code{FALSE} to disable or use \code{list(p = 0.9, n = 50)} to customize threshold values.}

\item{widget_char}{Widget to use for \code{character} variables: \code{\link[shinyWidgets:pickerInput]{shinyWidgets::pickerInput()}}
or \code{\link[shiny:selectInput]{shiny::selectInput()}} (default).}

\item{widget_num}{Widget to use for \code{numeric} variables: \code{\link[shinyWidgets:numericRangeInput]{shinyWidgets::numericRangeInput()}}
or \code{\link[shiny:sliderInput]{shiny::sliderInput()}} (default).}

\item{widget_date}{Widget to use for \code{date/time} variables: \code{\link[shiny:dateRangeInput]{shiny::dateRangeInput()}}
or \code{\link[shiny:sliderInput]{shiny::sliderInput()}} (default).}

\item{label_na}{Label for missing value widget.}

\item{value_na}{Default value for all NA's filters.}
}
\value{
\itemize{
\item UI: HTML tags that can be included in shiny's UI
\item Server: a \code{list} with four slots:
\itemize{
\item \strong{filtered}: a \code{reactive} function returning the data filtered.
\item \strong{code}: a \code{reactive} function returning the dplyr pipeline to filter data.
\item \strong{expr}: a \code{reactive} function returning an expression to filter data.
\item \strong{values}: a \code{reactive} function returning a named list of variables and filter values.
}
}
}
\description{
Module generate inputs to filter \code{data.frame} according column's type.
Code to reproduce the filter is returned as an expression with filtered data.
}
