% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filteredReactiveFacileDataStore-module.R
\name{filteredReactiveFacileDataStore}
\alias{filteredReactiveFacileDataStore}
\alias{filteredReactiveFacileDataStoreUI}
\alias{fds.FilteredReactiveFacileDataStore}
\title{Composition of modules to provide a filtering schema over a FaclieDataStore.}
\usage{
filteredReactiveFacileDataStore(input, output, session, dataset,
  active_samples = NULL, user = Sys.getenv("USER"), ...)

filteredReactiveFacileDataStoreUI(id, ...)

\method{fds}{FilteredReactiveFacileDataStore}(x, ...)
}
\description{
We compose the \code{\link[=reactiveFacileDataStore]{reactiveFacileDataStore()}} and \code{\link[=facileSampleFilter]{facileSampleFilter()}} modules
to make a super-module. The return objects from these two modules are
provided as "top-level" elements of the list returned by this module.
}
