% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filteredReactiveFacileDataStore.R,
%   R/filteredReactiveFacileDataStore_v1-module.R
\name{filteredReactiveFacileDataStoreUI}
\alias{filteredReactiveFacileDataStoreUI}
\alias{singleFilteredReactiveFacileDataStoreUI}
\alias{filteredReactiveFacileDataStore_v1}
\alias{filteredReactiveFacileDataStore_v1UI}
\title{Composition of modules to provide a filtering schema over a FaclieDataStore.}
\usage{
filteredReactiveFacileDataStoreUI(id, nfilters = 1, ...)

singleFilteredReactiveFacileDataStoreUI(id, nfilters = 1, ...)

filteredReactiveFacileDataStore_v1(input, output, session, dataset,
  active_samples = NULL, user = Sys.getenv("USER"), ...)

filteredReactiveFacileDataStore_v1UI(id, ...)
}
\description{
We compose the \code{\link[=reactiveFacileDataStore]{reactiveFacileDataStore()}} and \code{\link[=facileSampleFilter]{facileSampleFilter()}} modules
to make a super-module. The return objects from these two modules are
provided as "top-level" elements of the list returned by this module.
}
