% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamodFacileDataStoreServer.R
\name{filterSamplesTableServer}
\alias{filterSamplesTableServer}
\title{Show the filtered samples table for a ReactiveFacileDataSet}
\usage{
filterSamplesTableServer(id, dfds, ..., debug = FALSE)
}
\description{
Show the filtered samples table for a ReactiveFacileDataSet
}
