% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{from_fds}
\alias{from_fds}
\title{Check that the data of a module comes from the expected fds.}
\usage{
from_fds(x, rfds, ...)
}
\arguments{
\item{x}{the moduleServer}

\item{fds}{the ReactivFacileDataStore to test}
}
\description{
Check that the data of a module comes from the expected fds.
}
