% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fscatterPlot-modules.R
\name{fscatterPlotServer}
\alias{fscatterPlotServer}
\title{Module server for scatter plots}
\usage{
fscatterPlotServer(
  id,
  rfds,
  ...,
  gdb = shiny::reactive(NULL),
  ndim = 3,
  x = NULL,
  y = NULL,
  z = NULL,
  event_source = session$ns("selection"),
  .reactive = TRUE,
  debug = FALSE
)
}
\description{
Module server for scatter plots
}
