% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launch}
\alias{launch}
\title{Convenience function to launch the current version of the app}
\usage{
launch(config = NULL, user = Sys.getenv("USER"), ...,
  .log_level_shine = "warn", .log_level_data = "warn",
  .log_level_analysis = "warn")
}
\description{
Convenience function to launch the current version of the app
}
\seealso{
Other workbench functions: \code{\link{load_config}},
  \code{\link{staticbenchplus}}
}
\concept{workbench functions}
