% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_config}
\alias{load_config}
\title{Returns a config object for the workbench}
\usage{
load_config(config = NULL)
}
\description{
Returns a config object for the workbench
}
\seealso{
Other workbench functions: \code{\link{launch}},
  \code{\link{staticbenchplus}}
}
\concept{workbench functions}
