% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeReactiveTrigger.R
\name{makeReactiveTrigger}
\alias{makeReactiveTrigger}
\title{Create a reactive trigger to listen to and propagate application events}
\usage{
makeReactiveTrigger()
}
\value{
a list with \code{$depend()} and \code{$trigger()} to be used in reactive
expressions across your app.
}
\description{
A reactive trigger can be used when you want to be able to explicitly trigger
a reactive expression. You can think of it as being similar to an action
button, except instead of clicking on a button to trigger an expression, you
can programatically cause the trigger.
}
\details{
This concept and code was created by Joe Cheng, and you can find more
references about it at these links:
\itemize{
\item https://github.com/daattali/advanced-shiny/tree/master/reactive-trigger,
the documentation in this funciton is largely taken from this resource.
\item https://github.com/MangoTheCat/dynshiny#triggering-ui-changes
}

Reactive triggers are instantiated like so:
\code{myTrigger <- makeReactiveTrigger()}. To use it, put \code{myTrigger$depend()} in
any reactive code that should re-run when the trigger is fired, and call
\code{myTrigger$trigger()} to set off the trigger.
}
