% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-faciledatasets-directory.R
\name{parse_faciledatasets_directory}
\alias{parse_faciledatasets_directory}
\title{Lists the FacileDataSet-looking directories in a parent directory.}
\usage{
parse_faciledatasets_directory(datadir, metafn = NULL, ...)
}
\arguments{
\item{datadir}{the parent directory that holds the FacileDataSet directories}

\item{metafn}{an optional argument that identifies the yaml file that holds
metadata about the datasets in \code{datadir}}
}
\value{
a tibble of information for the FacileDataSets in the \code{datadir}
directory.
}
\description{
Currently the directories under \code{datadir} are recognized as a FacileDataSet
root if they have a \code{meta.yaml} file with top-level \code{name} and \code{organism}
attributes.
}
