% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantitativeAssayDataSelect.R
\name{quantitativeAssayDataSelect}
\alias{quantitativeAssayDataSelect}
\alias{quantitativeAssayDataSelectUI}
\title{Retrieves quantitative assay data from a FacileDataStore}
\usage{
quantitativeAssayDataSelect(input, output, session, rfds, ...,
  .exclude = NULL, .reactive = TRUE)

quantitativeAssayDataSelectUI(id, multiple = TRUE, ...)
}
\arguments{
\item{multiple, ...}{passed into the \code{"features"} \code{selectizeInput}}
}
\description{
We will compose this with the quantitativeSampleCovariate module to create
a super quantitativeTraitSelect, which enables the user to select assay
or quantitative covariates from a FacileDataStore
}
