% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantitativeTraitSelect-module.R
\name{quantitativeTraitSelect}
\alias{quantitativeTraitSelect}
\title{Provides a widget to select quantitative traits from samples}
\usage{
quantitativeTraitSelect(input, output, session, rfds, ...,
  .exclude = NULL, .reactive = TRUE)
}
\description{
This module will pull quantitative traits from a \code{FacileDataStore}.
}
\details{
We assume that there are two sources of quanitative traits:
\enumerate{
\item From assays (usually, dense matrix types of things (rnaseq)); or
\item A quantitative sample covariate (RIN score, for instance)
}

When type is "assay", multiple entries can be used. In these cases these
entries might be treated as a geneset, or individual scores.

Quantitative traits can only be selected one at a time. Further, let's assume
they are normalized for now,
}
