% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactiveFacileDatastoreSelect-module.R
\name{reactiveFacileDataStoreSelect}
\alias{reactiveFacileDataStoreSelect}
\title{Provides a selection widget to pick from a collection of FacileDataStores}
\usage{
reactiveFacileDataStoreSelect(input, output, session, fdslist, ...)
}
\arguments{
\item{fdslist}{a named list of FacileDataStore objects}
}
\description{
TODO: Final selection should be upload(?)
}
