% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactiveFacileDataStore-module.R
\name{reactives}
\alias{reactives}
\title{Extracts the "reactive" list from a FacileDataStore}
\usage{
reactives(x, name = NULL, ...)
}
\description{
An S3 ReactiveFacileDataStore (\code{rfds}) stores the reactives list in
\code{rfds[[".reactive."]]}. An S4 ReactiveFacileDataStore (ie. a FacileDataStore
over a \code{SingleCellExperiment}) will store this somehwere else.
}
