% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{reset}
\alias{reset}
\title{Resets the state of a module}
\usage{
reset(x, ...)
}
\arguments{
\item{x}{The facile module to reset}
}
\description{
I'm imagining we will want this when some catastrophic error gets thrown
during interaction with a shiny module, and we want to reset its original
state of zen.
}
\details{
One time this might happen is within an interactive analysis where the
\code{\link[=active_samples]{active_samples()}} change from under a module's feet, and some samples
or covariates that once existed are now gone, then: boom.

TODO: Ubiquitously implement \code{reset()} functionality.
}
