% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{shine}
\alias{shine}
\title{Workhorse function to invoke shiny gadgets for FacileData}
\usage{
shine(x, ...)
}
\arguments{
\item{x}{A "facile" object, like a \code{FacileAnalysisResult}.}
}
\value{
An \code{FacileGadgetResult} object that encapsulates the "end product" of
the exploration and interaction over \code{x} from within the gadget.
}
\description{
Provides \code{selection}, and other getter functions that will enable
reproducibility of the results explored and "finalized" within the shiny
gadget to be recovered within the R workspace.
}
