% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticbenchplus.R
\name{staticbenchplus}
\alias{staticbenchplus}
\title{Launch a workbench shiny app}
\usage{
staticbenchplus(config = NULL, user = Sys.getenv("USER"), ...)
}
\arguments{
\item{user}{the user accessing it (used for "custom_key" stuff in annotations)}

\item{x}{The FacileDataStore}
}
\value{
shiny app
}
\description{
Launch a workbench shiny app
}
\seealso{
Other workbench functions: \code{\link{launch}},
  \code{\link{load_config}}
}
\concept{workbench functions}
