% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toq}
\alias{toq}
\title{Convenience funciton for tictoc::tic(..., quiet = TRUE)}
\usage{
toq(log = FALSE, quiet = TRUE, func.toc = tictoc::toc.outmsg, ..., digits = 3)
}
\description{
Convenience funciton for tictoc::tic(..., quiet = TRUE)
}
