% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facileSampleFilter-module.R
\name{update_filters}
\alias{update_filters}
\title{Updates the filters used to set the active dataset}
\usage{
update_filters(x, filters, ...)
}
\description{
Updates the filters used to set the active dataset
}
