% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactiveFacileDataStore-module.R
\name{update_reactive_samples}
\alias{update_reactive_samples}
\alias{update_reactive_covariates}
\title{Update reactive samples and annotations}
\usage{
update_reactive_samples(x, actives_samples, criterion = NULL, ...)

update_reactive_covariates(x, covariates, namespace, ...)
}
\description{
We want to enable external actors to update the state of a reactive facile
datasetore.
}
