% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{update_selected}
\alias{update_selected}
\title{Updates a selected item from a module.}
\usage{
update_selected(x, covariate, ...)
}
\description{
This can apply to an item selected in a dropdown, ie. a selected covariate
in the categoricalSampleCovariate module, or perhaps a selected set of data
brushed somewhere(?)
}
