% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assayFeatureSelect-module.R,
%   R/assayFeatureSelect2-module.R
\name{assayFeatureSelect}
\alias{assayFeatureSelect}
\alias{assayFeatureSelectUI}
\alias{assayFeatureSelect2UI}
\title{Retrieves assay features from a FacileDataStore for a given assay type.}
\usage{
assayFeatureSelect(
  input,
  output,
  session,
  rfds,
  gdb = NULL,
  ...,
  .exclude = NULL,
  .reactive = TRUE
)

assayFeatureSelectUI(id, label = NULL, multiple = TRUE, ...)

assayFeatureSelect2UI(id, label = NULL, multiple = TRUE, ...)
}
\arguments{
\item{multiple, ...}{passed into the \code{"features"} \code{selectizeInput}}
}
\value{
a list with the following elements:
\itemize{
\item \code{assay_info}: one row assay,assay_type,feature_type tibble
\item \code{features}: n-row feature_info_tbl() like tbl enumerating the assay
features selected in this module
\item \code{features_all}: a tibble of all of the fdatures of this \code{feature_type}
(we can possible axe this, but ...)
\item \code{label}: a "human readable" summary of the features selected within this
module
\item \code{name}: a "computerfriendly" version of \code{label}
}
}
\description{
Retrieves assay features from a FacileDataStore for a given assay type.
}
