% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categoricalAestheticMap-module.R
\name{categoricalAestheticMap}
\alias{categoricalAestheticMap}
\alias{categoricalAestheticMapUI}
\title{Provides re-usable a widget bar to map data to aesthetics.}
\usage{
categoricalAestheticMap(input, output, session, rfds, color = FALSE,
  facet = FALSE, group = FALSE, hover = FALSE, shape = FALSE, ...,
  .with_none = TRUE, .exclude = NULL, .reactive = TRUE)

categoricalAestheticMapUI(id, color = FALSE, facet = FALSE,
  group = FALSE, hover = FALSE, shape = FALSE, horizontal = TRUE,
  ...)
}
\description{
Provides re-usable a widget bar to map data to aesthetics.
}
