% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categoricalSampleCovariate-modules.R
\name{categoricalSampleCovariateLevels}
\alias{categoricalSampleCovariateLevels}
\title{Use this with categoricalSampleCovariateSelect to enumerate its levels.}
\usage{
categoricalSampleCovariateLevels(input, output, session, rfds, covariate,
  ..., .exclude = NULL, .reactive = TRUE, debug = FALSE)
}
\arguments{
\item{covaraite}{the \code{categoricalSampleCovariateSelect} module.}
}
\description{
Use this with categoricalSampleCovariateSelect to enumerate its levels.
}
