% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categoricalSampleCovariate-modules.R
\name{categoricalSampleCovariateSelect}
\alias{categoricalSampleCovariateSelect}
\title{A selectInput that provides the categorical covariates over a FacileDataStore}
\usage{
categoricalSampleCovariateSelect(input, output, session, rfds, ...,
  .with_none = TRUE, .exclude = NULL, .reactive = TRUE)
}
\arguments{
\item{rfds}{A \code{ReactiveFacileDataStore}}

\item{.exclude}{a covariate tibble that incluees covariate,value pairs that
should not be used in this covariate selector. This can be a reactive
tibble if you want to be able to update the things one can't select for.}

\item{.reactive}{If \code{TRUE} (default), the module will constantly update the
covariates made available based on changes to the
\code{rfds[["active_covariates]]()} tibble. Otherwise, the covariates made
availabe (or .excluded) are fixed on module creation.}
}
\description{
Use in conjunction with a categoricalSampleCovariateLevels module.
}
\details{
Why am I using all of these seriously tortured names?
}
