% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categoricalSampleCovariate-modules.R
\name{categoricalSampleCovariateSelect}
\alias{categoricalSampleCovariateSelect}
\title{A categoricalSampleCovariate selector}
\usage{
categoricalSampleCovariateSelect(input, output, session, rfds,
  universe = NULL, ..., .with_none = TRUE, .exclude = NULL,
  .reactive = TRUE, debug = FALSE)
}
\description{
This is the second version of this module, which was updated to allow for a
\code{universe} argument, which specifies a subset of samples from \code{rfds} to get
covariate data from. The original implementation always retrieved covariates
from \code{active_samples(rfds)}. If \code{universe = NULL}, original behavior is
kept.
}
