% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categoricalSampleCovrariateSelect-module.R
\name{categoricalSampleCovariateSelectServer}
\alias{categoricalSampleCovariateSelectServer}
\title{A categoricalSampleCovariate selector}
\usage{
categoricalSampleCovariateSelectServer(
  id,
  rfds,
  include1 = TRUE,
  default_covariate = NULL,
  ...,
  with_none = TRUE,
  exclude = reactive(NULL),
  ignoreNULL = with_none,
  debug = FALSE
)
}
\arguments{
\item{default_covariate}{the name of a covariate to use as the default
one that is selected here, if available}

\item{a}{reactive character vector, which specifies the covariates to ignore
or a tibble with "variable" (and optional "value") columns. If a character
vector, or tibble with just has a "variable" column are provide, then
the variable names enumerated there will not be included in this dropdown.
A tibble with "variable" and "value" columns can be used so that only
specific levels of a covariate are ignored.}
}
\description{
This is the second version of this module, which was updated to allow for a
\code{universe} argument, which specifies a subset of samples from \code{rfds} to get
covariate data from. The original implementation always retrieved covariates
from \code{active_samples(rfds)}. If \code{universe = NULL}, original behavior is
kept.
}
\section{Excluding covariates}{


TODO: Let's talk about how to exclude covariates and their levels, and point
users to \code{\link[=update_exclude]{update_exclude()}}
}

