% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facileDataStoreServer.R
\name{FacileDataStoreServer}
\alias{FacileDataStoreServer}
\title{Crate a facileDataStoreServer module form whatevrrrrrrrrr}
\usage{
FacileDataStoreServer(id, x, samples_subset = NULL, ...)
}
\arguments{
\item{id}{the id of the server module}

\item{x}{A thing, this will be passed into the reactive
\verb{facileDataStoreServer,x} parameter.}

\item{...}{things to pass into the \code{facileDataStoreServer(...)}}
}
\value{
a \code{\link[=facileDataStoreServer]{facileDataStoreServer()}} module
}
\description{
Crate a facileDataStoreServer module form whatevrrrrrrrrr
}
