% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facileScatterPlot-module.R
\name{facileScatterPlot}
\alias{facileScatterPlot}
\alias{facileScatterPlotUI}
\title{An interactive scatterplot module}
\usage{
facileScatterPlot(input, output, session, rfds, ..., .reactive = TRUE)

facileScatterPlotUI(id, ...)
}
\description{
An interactive scatterplot module
}
