% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filteredReactiveFacileDataStore.R
\name{filteredReactiveFacileDataStoreServer}
\alias{filteredReactiveFacileDataStoreServer}
\alias{filteredReactiveFacileDataStore}
\title{Composition of modules to provide a filtering schema over a FaclieDataStore.}
\usage{
filteredReactiveFacileDataStoreServer(
  id,
  path,
  user = Sys.getenv("USER"),
  ...,
  debug = FALSE
)

filteredReactiveFacileDataStore(
  input,
  output,
  session,
  path,
  user = Sys.getenv("USER"),
  ...,
  debug = FALSE
)
}
\description{
We compose the \code{\link[=reactiveFacileDataStore]{reactiveFacileDataStore()}} and a list of
\code{\link[=facileSampleFilter]{facileSampleFilter()}} modules to make a super-module.

We compose the \code{\link[=reactiveFacileDataStore]{reactiveFacileDataStore()}} and a list of
\code{\link[=facileSampleFilter]{facileSampleFilter()}} modules to make a super-module.
}
