% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facileSampleFilter-module.R,
%   R/filteredReactiveFacileDataStore_v1-module.R
\name{filters}
\alias{filters}
\title{Returns a representaiton of the filters used to specify the active dataset}
\usage{
filters(x, ...)

filters(x, ...)
}
\description{
Returns a representaiton of the filters used to specify the active dataset

Returns a representaiton of the filters used to specify the active dataset
}
