% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{from_fds}
\alias{from_fds}
\title{Check that the data of a module comes from the expected fds.}
\usage{
from_fds(x, rfds, ...)
}
\arguments{
\item{x}{the moduleServer}

\item{fds}{the ReactivFacileDataStore to test}
}
\description{
This was a temporary (2 day) fix. I think the \code{module$in_sync()} paradigm
is a better one to ensure that the data that \code{module} is returning is still
the same internal \code{rfds} that it came from.
}
