% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantitativeAssayDataSelect.R
\name{quantitativeAssayDataSelect}
\alias{quantitativeAssayDataSelect}
\alias{quantitativeAssayDataSelectUI}
\title{Retrieves quantitative assay data from a FacileDataStore}
\usage{
quantitativeAssayDataSelect(input, output, session, rfds, ...,
  .exclude = NULL, .reactive = TRUE)

quantitativeAssayDataSelectUI(id, multiple = TRUE, ...)
}
\arguments{
\item{multiple, ...}{passed into the \code{"features"} \code{selectizeInput}}
}
\value{
a list with the following elements:
\itemize{
\item \code{assay_info}: one row assay,assay_type,feature_type tibble
\item \code{features}: n-row feature_info_tbl() like tbl enumerating the assay
features selected in this module
\item \code{features_all}: a tibble of all of the fdatures of this \code{feature_type}
(we can possible axe this, but ...)
\item \code{label}: a "human readable" summary of the features selected within this
module
\item \code{name}: a "computerfriendly" version of \code{label}
}
}
\description{
We will compose this with the quantitativeSampleCovariate module to create
a super quantitativeTraitSelect, which enables the user to select assay
or quantitative covariates from a FacileDataStore
}
