% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R, R/reactiveFacileDataStore-module.R,
%   R/reactiveFacileDataStore_v1-module.R
\name{active_assays}
\alias{active_assays}
\alias{active_covariates}
\alias{active_samples}
\alias{reactiveFacileDataStore}
\alias{reactiveFacileDataStore_v1UI}
\alias{active_assays.ReactiveFacileDataStore_v1}
\title{Creates a (swappable) ReactiveFacileDataSet}
\usage{
active_assays(x, ...)

active_covariates(x, ...)

active_samples(x, ...)

reactiveFacileDataStore(intput, output, session, path,
  user = Sys.getenv("USER"), ..., restrict_samples. = NULL)

reactiveFacileDataStore_v1UI(id, ...)

\method{active_assays}{ReactiveFacileDataStore_v1}(x, ...)
}
\arguments{
\item{path}{A "reactive string" that points to the url/directory to
instantiate the ReactiveFacileDataStore from.}
}
\description{
This module creates a reactive container for a FacileDataStore, by providing
a (reactive) \code{path} parameter that points to an accessible FacileDataStore.
The reactivity of this input parameter allows the datastore to be swapped out
"live".
}
\details{
To restrict the set of samples "in play" for the datastore, for instance when
invoking an interactive gadget over these samples, the analyst (or gadget)
can do, you can pass in a sample descriptor into the \code{restrict_samples.}
parameter. This parameter isn't reactive and isn't really meant to be
used outside of the simple context of pulling up a subset of a
FacileDataStore for interrogation via a gadget.

To allow interactive expansion and collapse of samples in play while
interactively working with a FacileDataStore with shiny, you should use
the \code{\link[=filteredReactiveDataStore]{filteredReactiveDataStore()}} module.
}
