% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleFilterList-module.R
\name{sampleFilterList}
\alias{sampleFilterList}
\title{An expandable (and shrinkable) list of sampleFilter modules.}
\usage{
sampleFilterList(input, output, session, rfds, ..., debug = FALSE)
}
\description{
Implementation largely inspired from the \code{filter_module.R} file in Joe
Cheng's rpharma-demo shiny app and a post from Barret Schloerke in the
RStudio community forums (cf. the "See Also" section).
}
\seealso{
\href{https://github.com/jcheng5/rpharma-demo}{Joe Cheng's rpharm-demo app},
\href{https://community.rstudio.com/t/22617/5}{Barret's Post}
}
