% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactiveFacileDataStore_v1-module.R
\name{update_reactive_samples}
\alias{update_reactive_samples}
\alias{update_reactive_covariates}
\alias{update_reactive_samples.ReactiveFacileDataStore_v1}
\alias{update_reactive_covariates.ReactiveFacileDataStore_v1}
\title{Update reactive samples and annotations}
\usage{
update_reactive_samples(x, actives_samples, criterion = NULL, ...)

update_reactive_covariates(x, covariates, namespace, ...)

\method{update_reactive_samples}{ReactiveFacileDataStore_v1}(x,
  active_samples, criterion = NULL, ...)

\method{update_reactive_covariates}{ReactiveFacileDataStore_v1}(rfds,
  covariates, namespace, ...)
}
\description{
We want to enable external actors to update the state of a reactive facile
datasetore.
}
\section{Reactive Samples}{


A facile analysis session is perfomed on some subset of the samples
provided by a \code{FacileDataStore}. The samples under scrutiny can change as
the user dips into and out of different analyses. These are defined
interactively by the user via the use of the \code{facileSampleFilter} module(s)
}

\section{Reactive Covariates}{


Updates the ephemeral covariates during an analysis.

In addition to the sample covariates in the FacileDataStore, covariates
can be added to the samples by different modules via user interaction and
brushing during an analysis session.
}

\examples{
fds <- FacileData::exampleFacileDataSet()
covs <- samples(fds) \%>\%
  collect() \%>\%
  sample_n(20) \%>\%
  mutate(variable = "test_anno",
         value = sample(c("a", "b"), 20, replace = TRUE),
         class = "categorical",
         type = "ephemeral",
         date_entered = as.integer(Sys.time()))
ufds <- update_reactive_covariates(fds, covs, "some_module")
}
